[FEATURE] [ANNOTATION-13] Added support to toggle password visibility.
The functionality is complete. But the design is yet to be implemented. There is an error occurring at the time of font-awesome react package installation. This will be taken care of shortly.

Currently, this feature only accessible from the re-enter password field from signup page through a button. Once the design is fixed, this feature will be replicated in applicable areas of signup and sign-in.